+++ AwtadStudy/Universities/Courses/Course.cs
@@ -1,4 +1,4 @@
-﻿namespace AwtadStudy.Universities;
+﻿namespace AwtadStudy.Universities.Courses;
 
 public sealed record Course(string ID,
                             string Name,
