+++ AwtadStudy/Universities/Courses/CourseGroup.cs
@@ -1,0 +1,6 @@
+﻿namespace AwtadStudy.Universities.Courses;
+
+public sealed record CourseGroup(string GroupID,
+                                 string Lecturer,
+                                 IEnumerable<LectureInfo> Lectures,
+                                 IEnumerable<ExamInfo> Exams);
