+++ AwtadStudy/Universities/Courses/ExamInfo.cs
@@ -1,0 +1,3 @@
+﻿namespace AwtadStudy.Universities.Courses;
+
+public sealed record ExamInfo(DateOnly Date, Moed Moed, ExamType Type);
