+++ AwtadStudy/Universities/Courses/ExamType.cs
@@ -1,0 +1,7 @@
+﻿namespace AwtadStudy.Universities.Courses;
+
+public enum ExamType
+{
+    Final,
+    Midterm,
+}
