+++ AwtadStudy/Universities/Courses/LectureInfo.cs
@@ -1,0 +1,3 @@
+﻿namespace AwtadStudy.Universities.Courses;
+
+public sealed record LectureInfo(DayOfWeek Day, TimeOnly StartTime, TimeOnly EndTime, string Place);
