+++ AwtadStudy/Universities/Courses/Moed.cs
@@ -1,0 +1,8 @@
+﻿namespace AwtadStudy.Universities.Courses;
+
+public enum Moed
+{
+    A,
+    B,
+    Special,
+}
