+++ AwtadStudy/Universities/University.cs
@@ -1,0 +1,9 @@
+﻿namespace AwtadStudy.Universities;
+
+public enum University
+{
+    TelAviv,
+    Technion,
+    BarIlan,
+    Haifa,
+}
